requirements.txt
functions-framework
google-cloud-firestore
google-cloud-storage
werkzeug